RCB-5-Create recipe controller
