+++ recipe-box-backend/src/main/java/com/endava/recipebox/controller/RecipeController.java
@@ -1,14 +1,17 @@
-package com.endava.recipebox.recipe;
+package com.endava.recipebox.controller;
 
+import com.endava.recipebox.model.MealType;
+import com.endava.recipebox.model.Recipe;
+import com.endava.recipebox.service.RecipeService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.stream.Collectors;
+import java.util.List;
+
 
 @RestController
 @RequestMapping(path = "/recipe_box", produces = "application/json")
@@ -28,12 +31,12 @@ public class RecipeController {
 
     @GetMapping("/publicRecipes/{name}")
     public ResponseEntity<List<Recipe>> getAllPublicRecipesByName(@PathVariable(name = "recipeName") String name) {
-        return ResponseEntity.ok(recipeService.findRecipeByName(name));
+        return ResponseEntity.ok(recipeService.getAllPublicRecipesByName(name));
     }
 
     @GetMapping("/publicRecipes/{mealType}")
-    public ResponseEntity<List<Recipe>> getAllPublicRecipesByType(@PathVariable(name = "mealType") String mealType) {
-        return ResponseEntity.ok(recipeService.findRecipeByType(mealType));
+    public ResponseEntity<List<Recipe>> getAllPublicRecipesByType(@PathVariable(name = "mealType") MealType mealType) {
+        return ResponseEntity.ok(recipeService.getAllPublicRecipesByType(mealType));
 
     }
