+++ recipe-box-backend/src/main/java/com/endava/recipebox/repository/RecipeRepository.java
@@ -1,8 +1,11 @@
-package com.endava.recipebox.recipe;
+package com.endava.recipebox.repository;
 
+import com.endava.recipebox.model.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
 public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+
 }
